rename module `color` to `colored`
