+++ src/colored/mod.rs
